Removed library loading. This was done to simplify the code going forward. This will cost the already somewhat broken backwards-compatibility though.
