+++ GERTe/GEDS/NetString.cpp
