+++ GERTe/GEDS/NetString.h
