+++ GERTe/GEDS/Ports.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #ifdef _WIN32
+#define WIN32_LEAN_AND_MEAN
 #include <WinSock2.h>
 #pragma comment(lib, "Ws2_32.lib")
 #else
