+++ GERTe/GEDS/libLoad.h
@@ -1,8 +1,6 @@
 #pragma once
-#include "Status.h"
 #include "Version.h"
 using namespace std;
 
-Status loadLibs();
 Version* getVersion(unsigned char);
 unsigned char highestVersion();
